moving shared code out of helm and upgrade tests
