+++ test/kube2e/upgrade/helm/helm_test.go
@@ -4,32 +4,20 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"io/ioutil"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"strings"
-	"text/template"
-	"time"
-
 	"github.com/ghodss/yaml"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	gatewayv1kube "github.com/solo-io/gloo/projects/gateway/pkg/api/v1/kube/client/clientset/versioned/typed/gateway.solo.io/v1"
-	"github.com/solo-io/gloo/projects/gloo/cli/pkg/cmd/version"
-	"github.com/solo-io/gloo/projects/gloo/cli/pkg/helpers"
 	"github.com/solo-io/gloo/projects/gloo/pkg/defaults"
 	"github.com/solo-io/gloo/test/kube2e"
+	"github.com/solo-io/gloo/test/kube2e/upgrade"
 	exec_utils "github.com/solo-io/go-utils/testutils/exec"
 	"github.com/solo-io/k8s-utils/kubeutils"
 	"github.com/solo-io/k8s-utils/testutils/helper"
 	"github.com/solo-io/skv2/codegen/util"
-	"github.com/solo-io/solo-kit/pkg/api/v1/clients"
 	"github.com/solo-io/solo-kit/pkg/code-generator/schemagen"
 	admission_v1 "k8s.io/api/admissionregistration/v1"
-	corev1 "k8s.io/api/core/v1"
 	v1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -37,6 +25,9 @@ import (
 	admission_v1_types "k8s.io/client-go/kubernetes/typed/admissionregistration/v1"
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 	"k8s.io/client-go/rest"
+	"os"
+	"path/filepath"
+	"text/template"
 )
 
 // now that we run CI on a kube 1.22 cluster, we must ensure that we install versions of gloo with v1 CRDs
@@ -93,7 +84,7 @@ var _ = Describe("Kube2e: helm", func() {
 	})
 
 	AfterEach(func() {
-		uninstallGloo(testHelper, ctx, cancel)
+		upgrade.UninstallGloo(testHelper, ctx, cancel)
 	})
 
 	Context("upgrades", func() {
@@ -101,93 +92,12 @@ var _ = Describe("Kube2e: helm", func() {
 			fromRelease = earliestVersionWithV1CRDs
 		})
 
-		It("uses helm to upgrade to this gloo version without errors", func() {
-
-			By("should start with gloo version 1.9.0")
-			Expect(getGlooServerVersion(ctx, testHelper.InstallNamespace)).To(Equal(earliestVersionWithV1CRDs))
-
-			// upgrade to the gloo version being tested
-			upgradeGloo(testHelper, chartUri, crdDir, fromRelease, strictValidation, nil)
-
-			By("should have upgraded to the gloo version being tested")
-			Expect(getGlooServerVersion(ctx, testHelper.InstallNamespace)).To(Equal(testHelper.ChartVersion()))
-		})
-
-		It("uses helm to update the settings without errors", func() {
-
-			By("should start with the settings.invalidConfigPolicy.invalidRouteResponseCode=404")
-			client := helpers.MustSettingsClient(ctx)
-			settings, err := client.Read(testHelper.InstallNamespace, defaults.SettingsName, clients.ReadOpts{})
-			Expect(err).To(BeNil())
-			Expect(settings.GetGloo().GetInvalidConfigPolicy().GetInvalidRouteResponseCode()).To(Equal(uint32(404)))
-
-			upgradeGloo(testHelper, chartUri, crdDir, fromRelease, strictValidation, []string{
-				"--set", "settings.replaceInvalidRoutes=true",
-				"--set", "settings.invalidConfigPolicy.invalidRouteResponseCode=400",
-			})
-
-			By("should have updated to settings.invalidConfigPolicy.invalidRouteResponseCode=400")
-			settings, err = client.Read(testHelper.InstallNamespace, defaults.SettingsName, clients.ReadOpts{})
-			Expect(err).To(BeNil())
-			Expect(settings.GetGloo().GetInvalidConfigPolicy().GetInvalidRouteResponseCode()).To(Equal(uint32(400)))
-		})
-
-		It("uses helm to update the validationServerGrpcMaxSizeBytes without errors", func() {
-
-			// this is the default value from the 1.9.0 chart
-			By("should start with the gateway.validation.validationServerGrpcMaxSizeBytes=4000000 (4MB)")
-			client := helpers.MustSettingsClient(ctx)
-			settings, err := client.Read(testHelper.InstallNamespace, defaults.SettingsName, clients.ReadOpts{})
-			Expect(err).To(BeNil())
-			Expect(settings.GetGateway().GetValidation().GetValidationServerGrpcMaxSizeBytes().GetValue()).To(Equal(int32(4000000)))
-
-			upgradeGloo(testHelper, chartUri, crdDir, fromRelease, strictValidation, []string{
-				"--set", "gateway.validation.validationServerGrpcMaxSizeBytes=5000000",
-			})
-
-			By("should have updated to gateway.validation.validationServerGrpcMaxSizeBytes=5000000 (5MB)")
-			settings, err = client.Read(testHelper.InstallNamespace, defaults.SettingsName, clients.ReadOpts{})
-			Expect(err).To(BeNil())
-			Expect(settings.GetGateway().GetValidation().GetValidationServerGrpcMaxSizeBytes().GetValue()).To(Equal(int32(5000000)))
+		It("uses helm to upgrade to this gloo version and settings without errors", func() {
+			upgrade.UpdateSettingsWithoutErrors(ctx, testHelper, crdDir, earliestVersionWithV1CRDs, chartUri, strictValidation)
 		})
 
 		It("uses helm to add a second gateway-proxy in a separate namespace without errors", func() {
-			const externalNamespace = "other-ns"
-			requiredSettings := map[string]string{
-				"gatewayProxies.proxyExternal.disabled":              "false",
-				"gatewayProxies.proxyExternal.namespace":             externalNamespace,
-				"gatewayProxies.proxyExternal.service.type":          "NodePort",
-				"gatewayProxies.proxyExternal.service.httpPort":      "31500",
-				"gatewayProxies.proxyExternal.service.httpsPort":     "32500",
-				"gatewayProxies.proxyExternal.service.httpNodePort":  "31500",
-				"gatewayProxies.proxyExternal.service.httpsNodePort": "32500",
-			}
-
-			var settings []string
-			for key, val := range requiredSettings {
-				settings = append(settings, "--set")
-				settings = append(settings, strings.Join([]string{key, val}, "="))
-			}
-
-			runAndCleanCommand("kubectl", "create", "ns", externalNamespace)
-			defer runAndCleanCommand("kubectl", "delete", "ns", externalNamespace)
-
-			upgradeGloo(testHelper, chartUri, crdDir, fromRelease, strictValidation, settings)
-
-			// Ensures deployment is created for both default namespace and external one
-			// Note- name of external deployments is kebab-case of gatewayProxies NAME helm value
-			Eventually(func() (string, error) {
-				return exec_utils.RunCommandOutput(testHelper.RootDir, false,
-					"kubectl", "get", "deployment", "-A")
-			}, "10s", "1s").Should(
-				And(ContainSubstring("gateway-proxy"),
-					ContainSubstring("proxy-external")))
-
-			// Ensures service account is created for the external namespace
-			Eventually(func() (string, error) {
-				return exec_utils.RunCommandOutput(testHelper.RootDir, false,
-					"kubectl", "get", "serviceaccount", "-n", externalNamespace)
-			}, "10s", "1s").Should(ContainSubstring("gateway-proxy"))
+			upgrade.AddSecondGatewayProxySeparateNamespaceTest(testHelper, crdDir, chartUri, strictValidation)
 		})
 	})
 
@@ -206,25 +116,7 @@ var _ = Describe("Kube2e: helm", func() {
 		})
 
 		It("sets validation webhook caBundle on install and upgrade", func() {
-			webhookConfigClient := kubeClientset.AdmissionregistrationV1().ValidatingWebhookConfigurations()
-			secretClient := kubeClientset.CoreV1().Secrets(testHelper.InstallNamespace)
-
-			By("the webhook caBundle should be the same as the secret's root ca value")
-			webhookConfig, err := webhookConfigClient.Get(ctx, "gloo-gateway-validation-webhook-"+testHelper.InstallNamespace, metav1.GetOptions{})
-			Expect(err).NotTo(HaveOccurred())
-			secret, err := secretClient.Get(ctx, "gateway-validation-certs", metav1.GetOptions{})
-			Expect(err).NotTo(HaveOccurred())
-			Expect(webhookConfig.Webhooks[0].ClientConfig.CABundle).To(Equal(secret.Data[corev1.ServiceAccountRootCAKey]))
-
-			// do an upgrade
-			upgradeGloo(testHelper, chartUri, crdDir, fromRelease, strictValidation, nil)
-
-			By("the webhook caBundle and secret's root ca value should still match after upgrade")
-			webhookConfig, err = webhookConfigClient.Get(ctx, "gloo-gateway-validation-webhook-"+testHelper.InstallNamespace, metav1.GetOptions{})
-			Expect(err).NotTo(HaveOccurred())
-			secret, err = secretClient.Get(ctx, "gateway-validation-certs", metav1.GetOptions{})
-			Expect(err).NotTo(HaveOccurred())
-			Expect(webhookConfig.Webhooks[0].ClientConfig.CABundle).To(Equal(secret.Data[corev1.ServiceAccountRootCAKey]))
+			upgrade.UpdateValidationWebhookTests(ctx, crdDir, kubeClientset, testHelper, chartUri, false)
 		})
 
 		// Below are tests with different combinations of upgrades with failurePolicy=Ignore/Fail.
@@ -258,7 +150,7 @@ var _ = Describe("Kube2e: helm", func() {
 				if newFailurePolicy == admission_v1.Fail {
 					newStrictValue = true
 				}
-				upgradeGloo(testHelper, chartUri, crdDir, fromRelease, newStrictValue, []string{})
+				upgrade.GlooToBranchVersion(testHelper, chartUri, crdDir, newStrictValue, []string{})
 
 				By(fmt.Sprintf("should have updated to gateway.validation.failurePolicy=%v", newFailurePolicy))
 				webhookConfig, err = webhookConfigClient.Get(ctx, "gloo-gateway-validation-webhook-"+testHelper.InstallNamespace, metav1.GetOptions{})
@@ -363,7 +255,7 @@ var _ = Describe("Kube2e: helm", func() {
 			// Apply the Settings CRD to ensure it is the most up to date version
 			// this ensures that any new fields that have been added are included in the CRD validation schemas
 			settingsCrdFilePath := filepath.Join(crdDir, "gloo.solo.io_v1_Settings.yaml")
-			runAndCleanCommand("kubectl", "apply", "-f", settingsCrdFilePath)
+			upgrade.RunAndCleanCommand("kubectl", "apply", "-f", settingsCrdFilePath)
 		})
 
 		It("works using kubectl apply", func() {
@@ -391,20 +283,6 @@ var _ = Describe("Kube2e: helm", func() {
 	})
 })
 
-func getGlooServerVersion(ctx context.Context, namespace string) (v string) {
-	glooVersion, err := version.GetClientServerVersions(ctx, version.NewKube(namespace, ""))
-	Expect(err).To(BeNil())
-	Expect(len(glooVersion.GetServer())).To(Equal(1))
-	for _, container := range glooVersion.GetServer()[0].GetKubernetes().GetContainers() {
-		if v == "" {
-			v = container.Tag
-		} else {
-			Expect(container.Tag).To(Equal(v))
-		}
-	}
-	return v
-}
-
 func makeUnstructured(yam string) *unstructured.Unstructured {
 	jsn, err := yaml.YAMLToJSON([]byte(yam))
 	ExpectWithOffset(1, err).NotTo(HaveOccurred())
@@ -422,6 +300,8 @@ func makeUnstructuredFromTemplateFile(fixtureName string, values interface{}) *u
 	return makeUnstructured(b.String())
 }
 
+// Helm tests use the fromRelease parameter to determine the current
+// release version so cannot share code between other upgrade tests
 func installGloo(testHelper *helper.SoloTestHelper, chartUri string, fromRelease string, strictValidation bool) {
 	valueOverrideFile, cleanupFunc := kube2e.GetHelmValuesOverrideFile()
 	defer cleanupFunc()
@@ -429,7 +309,7 @@ func installGloo(testHelper *helper.SoloTestHelper, chartUri string, fromRelease
 	// construct helm args
 	var args = []string{"install", testHelper.HelmChartName}
 	if fromRelease != "" {
-		runAndCleanCommand("helm", "repo", "add", testHelper.HelmChartName,
+		upgrade.RunAndCleanCommand("helm", "repo", "add", testHelper.HelmChartName,
 			"https://storage.googleapis.com/solo-public-helm", "--force-update")
 		args = append(args, "gloo/gloo",
 			"--version", fmt.Sprintf("v%s", fromRelease))
@@ -440,128 +320,12 @@ func installGloo(testHelper *helper.SoloTestHelper, chartUri string, fromRelease
 		"--create-namespace",
 		"--values", valueOverrideFile)
 	if strictValidation {
-		args = append(args, strictValidationArgs...)
+		args = append(args, upgrade.StrictValidationArgs...)
 	}
 
 	fmt.Printf("running helm with args: %v\n", args)
-	runAndCleanCommand("helm", args...)
+	upgrade.RunAndCleanCommand("helm", args...)
 
 	// Check that everything is OK
-	checkGlooHealthy(testHelper)
-}
-
-// CRDs are applied to a cluster when performing a `helm install` operation
-// However, `helm upgrade` intentionally does not apply CRDs (https://helm.sh/docs/topics/charts/#limitations-on-crds)
-// Before performing the upgrade, we must manually apply any CRDs that were introduced since v1.9.0
-func upgradeCrds(testHelper *helper.SoloTestHelper, fromRelease string, crdDir string) {
-	// if we're just upgrading within the same release, no need to reapply crds
-	if fromRelease == "" {
-		return
-	}
-
-	// apply crds from the release we're upgrading to
-	runAndCleanCommand("kubectl", "apply", "-f", crdDir)
-	// allow some time for the new crds to take effect
-	time.Sleep(time.Second * 5)
-}
-
-func upgradeGloo(testHelper *helper.SoloTestHelper, chartUri string, crdDir string, fromRelease string, strictValidation bool, additionalArgs []string) {
-	upgradeCrds(testHelper, fromRelease, crdDir)
-
-	valueOverrideFile, cleanupFunc := getHelmUpgradeValuesOverrideFile()
-	defer cleanupFunc()
-
-	var args = []string{"upgrade", testHelper.HelmChartName, chartUri,
-		"-n", testHelper.InstallNamespace,
-		"--values", valueOverrideFile}
-	if strictValidation {
-		args = append(args, strictValidationArgs...)
-	}
-	args = append(args, additionalArgs...)
-
-	fmt.Printf("running helm with args: %v\n", args)
-	runAndCleanCommand("helm", args...)
-
-	// Check that everything is OK
-	checkGlooHealthy(testHelper)
-}
-
-func uninstallGloo(testHelper *helper.SoloTestHelper, ctx context.Context, cancel context.CancelFunc) {
-	Expect(testHelper).ToNot(BeNil())
-	err := testHelper.UninstallGlooAll()
-	Expect(err).NotTo(HaveOccurred())
-	_, err = kube2e.MustKubeClient().CoreV1().Namespaces().Get(ctx, testHelper.InstallNamespace, metav1.GetOptions{})
-	Expect(apierrors.IsNotFound(err)).To(BeTrue())
-	cancel()
-}
-
-func getHelmUpgradeValuesOverrideFile() (filename string, cleanup func()) {
-	values, err := ioutil.TempFile("", "values-*.yaml")
-	Expect(err).NotTo(HaveOccurred())
-
-	_, err = values.Write([]byte(`
-global:
-  image:
-    pullPolicy: IfNotPresent
-  glooRbac:
-    namespaced: true
-    nameSuffix: e2e-test-rbac-suffix
-settings:
-  singleNamespace: true
-  create: true
-  replaceInvalidRoutes: true
-gateway:
-  persistProxySpec: true
-gatewayProxies:
-  gatewayProxy:
-    healthyPanicThreshold: 0
-    gatewaySettings:
-      # the KEYVALUE action type was first available in v1.11.11 (within the v1.11.x branch); this is a sanity check to
-      # ensure we can upgrade without errors from an older version to a version with these new fields (i.e. we can set
-      # the new fields on the Gateway CR during the helm upgrade, and that it will pass validation)
-      customHttpGateway:
-        options:
-          dlp:
-            dlpRules:
-            - actions:
-              - actionType: KEYVALUE
-                keyValueAction:
-                  keyToMask: test
-                  name: test
-`))
-	Expect(err).NotTo(HaveOccurred())
-
-	err = values.Close()
-	Expect(err).NotTo(HaveOccurred())
-
-	return values.Name(), func() { _ = os.Remove(values.Name()) }
-}
-
-var strictValidationArgs = []string{
-	"--set", "gateway.validation.failurePolicy=Fail",
-	"--set", "gateway.validation.allowWarnings=false",
-	"--set", "gateway.validation.alwaysAcceptResources=false",
-}
-
-func runAndCleanCommand(name string, arg ...string) []byte {
-	cmd := exec.Command(name, arg...)
-	b, err := cmd.Output()
-	// for debugging in Cloud Build
-	if err != nil {
-		if v, ok := err.(*exec.ExitError); ok {
-			fmt.Println("ExitError: ", string(v.Stderr))
-		}
-	}
-	Expect(err).To(BeNil())
-	cmd.Process.Kill()
-	cmd.Process.Release()
-	return b
-}
-
-func checkGlooHealthy(testHelper *helper.SoloTestHelper) {
-	deploymentNames := []string{"gloo", "discovery", "gateway-proxy"}
-	for _, deploymentName := range deploymentNames {
-		runAndCleanCommand("kubectl", "rollout", "status", "deployment", "-n", testHelper.InstallNamespace, deploymentName)
-	}
-	kube2e.GlooctlCheckEventuallyHealthy(2, testHelper, "90s")
+	upgrade.CheckGlooOssHealthy(testHelper)
 }
