+++ test/kube2e/upgrade/upgrade_utils.go
@@ -3,11 +3,20 @@ package upgrade
 import (
 	"context"
 	"fmt"
+	. "github.com/onsi/gomega"
+	"github.com/solo-io/gloo/projects/gloo/cli/pkg/cmd/version"
+	"github.com/solo-io/gloo/test/kube2e"
+	"github.com/solo-io/k8s-utils/testutils/helper"
+	"io/ioutil"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"os"
+	"os/exec"
 	"path/filepath"
 	"runtime"
 	"sort"
 	"strings"
+	"time"
 
 	"github.com/google/go-github/v32/github"
 	errors "github.com/rotisserie/eris"
@@ -31,29 +40,28 @@ func (a ByVersion) Less(i, j int) bool {
 }
 func (a ByVersion) Swap(i, j int) { a[i], a[j] = a[j], a[i] }
 
-func GetUpgradeVersions(ctx context.Context) (lastMinorLatestPatchVersion *versionutils.Version, currentMinorLatestPatchVersion *versionutils.Version, err error) {
-	currentMinorLatestPatchVersion, curMinorErr := GetLastReleaseOfCurrentMinor()
+func GetUpgradeVersions(ctx context.Context, repoName string) (lastMinorLatestPatchVersion *versionutils.Version, currentMinorLatestPatchVersion *versionutils.Version, err error) {
+	currentMinorLatestPatchVersion, curMinorErr := GetLastReleaseOfCurrentMinor(repoName)
 	if curMinorErr != nil {
 		if curMinorErr.Error() != FirstReleaseError {
 			return nil, nil, curMinorErr
 		}
 	}
-	lastMinorLatestPatchVersion, lastMinorErr := GetLatestReleasedVersion(ctx, currentMinorLatestPatchVersion.Major, currentMinorLatestPatchVersion.Minor-1)
+	lastMinorLatestPatchVersion, lastMinorErr := GetLatestReleasedVersion(ctx, repoName, currentMinorLatestPatchVersion.Major, currentMinorLatestPatchVersion.Minor-1)
 	if lastMinorErr != nil {
 		return nil, nil, lastMinorErr
 	}
 	return lastMinorLatestPatchVersion, currentMinorLatestPatchVersion, curMinorErr
 }
 
-func GetLastReleaseOfCurrentMinor() (*versionutils.Version, error) {
-	repo_name := "gloo"                    // pull out to const
+func GetLastReleaseOfCurrentMinor(repoName string) (*versionutils.Version, error) { // pull out to const
 	_, filename, _, _ := runtime.Caller(0) //get info about what is calling the function
 	fmt.Printf(filename)
 	fParts := strings.Split(filename, string(os.PathSeparator))
 	splitIdx := 0
 	//we can end up in a situation where the path contains the repo_name twice when running in ci - keep going until we find the last use ex: /home/runner/work/gloo/gloo/test/kube2e/upgrade/junit.xml
 	for idx, dir := range fParts {
-		if dir == repo_name {
+		if dir == repoName {
 			splitIdx = idx
 		}
 	}
@@ -85,7 +93,7 @@ func GetLastReleaseOfCurrentMinor() (*versionutils.Version, error) {
 	return versions[len(versions)-2], nil
 }
 
-func GetLatestReleasedVersion(ctx context.Context, majorVersion, minorVersion int) (*versionutils.Version, error) {
+func GetLatestReleasedVersion(ctx context.Context, repoName string, majorVersion, minorVersion int) (*versionutils.Version, error) {
 	client, err := githubutils.GetClient(ctx)
 	if err != nil {
 		return nil, errors.Wrapf(err, "unable to create github client")
@@ -98,7 +106,7 @@ func GetLatestReleasedVersion(ctx context.Context, majorVersion, minorVersion in
 	for i := 0; i < 5; i++ {
 		// Get the next page of
 		listOpts := github.ListOptions{Page: i, PerPage: 10} // max per request
-		releases, _, err := client.Repositories.ListReleases(ctx, "solo-io", "gloo", &listOpts)
+		releases, _, err := client.Repositories.ListReleases(ctx, "solo-io", repoName, &listOpts)
 		if err != nil {
 			return nil, errors.Wrapf(err, "error listing releases")
 		}
@@ -122,6 +130,162 @@ func GetLatestReleasedVersion(ctx context.Context, majorVersion, minorVersion in
 			}
 		}
 	}
-
 	return nil, errors.Errorf("Could not find a recent release with version prefix: %s", versionPrefix)
+}
+
+func InstallGloo(testHelper *helper.SoloTestHelper, fromRelease string, strictValidation bool) {
+	valueOverrideFile, cleanupFunc := kube2e.GetHelmValuesOverrideFile()
+	defer cleanupFunc()
+
+	// construct helm args
+	var args = []string{"install", testHelper.HelmChartName}
+
+	RunAndCleanCommand("helm", "repo", "add", testHelper.HelmChartName,
+		"https://storage.googleapis.com/solo-public-helm", "--force-update")
+	args = append(args, "gloo/gloo",
+		"--version", fromRelease)
+
+	args = append(args, "-n", testHelper.InstallNamespace,
+		"--create-namespace",
+		"--values", valueOverrideFile)
+	if strictValidation {
+		args = append(args, StrictValidationArgs...)
+	}
+
+	fmt.Printf("running helm with args: %v\n", args)
+	RunAndCleanCommand("helm", args...)
+
+	// Check that everything is OK
+	CheckGlooOssHealthy(testHelper)
+}
+
+func UninstallGloo(testHelper *helper.SoloTestHelper, ctx context.Context, cancel context.CancelFunc) {
+	Expect(testHelper).ToNot(BeNil())
+	err := testHelper.UninstallGloo()
+	Expect(err).NotTo(HaveOccurred())
+	_, err = kube2e.MustKubeClient().CoreV1().Namespaces().Get(ctx, testHelper.InstallNamespace, metav1.GetOptions{})
+	Expect(apierrors.IsNotFound(err)).To(BeTrue())
+	cancel()
+}
+
+// CRDs are applied to a cluster when performing a `helm install` operation
+// However, `helm upgrade` intentionally does not apply CRDs (https://helm.sh/docs/topics/charts/#limitations-on-crds)
+// Before performing the upgrade, we must manually apply any CRDs that were introduced since v1.9.0
+func Crds(crdDir string) {
+	// apply crds from the release we're upgrading to
+	fmt.Printf("Upgrade crds: kubectl apply -f %s\n", crdDir)
+	RunAndCleanCommand("kubectl", "apply", "-f", crdDir)
+	// allow some time for the new crds to take effect
+	time.Sleep(time.Second * 10)
+}
+
+// upgrade the version of gloo to the branch version
+func GlooToBranchVersion(testHelper *helper.SoloTestHelper, chartUri string, crdDir string, strictValidation bool, additionalArgs []string) {
+	Crds(crdDir)
+
+	valueOverrideFile, cleanupFunc := GetHelmUpgradeValuesOverrideFile()
+	defer cleanupFunc()
+
+	var args = []string{"upgrade", testHelper.HelmChartName, chartUri,
+		"-n", testHelper.InstallNamespace,
+		"--values", valueOverrideFile}
+	if strictValidation {
+		args = append(args, StrictValidationArgs...)
+	}
+	args = append(args, additionalArgs...)
+
+	fmt.Printf("running helm with args: %v\n", args)
+	RunAndCleanCommand("helm", args...)
+
+	//Check that everything is OK
+	CheckGlooOssHealthy(testHelper)
+}
+
+func GetHelmUpgradeValuesOverrideFile() (filename string, cleanup func()) {
+	values, err := ioutil.TempFile("", "values-*.yaml")
+	Expect(err).NotTo(HaveOccurred())
+
+	_, err = values.Write([]byte(`
+global:
+  image:
+    pullPolicy: IfNotPresent
+  glooRbac:
+    namespaced: true
+    nameSuffix: e2e-test-rbac-suffix
+settings:
+  singleNamespace: true
+  create: true
+  replaceInvalidRoutes: true
+gateway:
+  persistProxySpec: true
+gatewayProxies:
+  gatewayProxy:
+    healthyPanicThreshold: 0
+    gatewaySettings:
+      # the KEYVALUE action type was first available in v1.11.11 (within the v1.11.x branch); this is a sanity check to
+      # ensure we can upgrade without errors from an older version to a version with these new fields (i.e. we can set
+      # the new fields on the Gateway CR during the helm upgrade, and that it will pass validation)
+      customHttpGateway:
+        options:
+          dlp:
+            dlpRules:
+            - actions:
+              - actionType: KEYVALUE
+                keyValueAction:
+                  keyToMask: test
+                  name: test
+`))
+	Expect(err).NotTo(HaveOccurred())
+
+	err = values.Close()
+	Expect(err).NotTo(HaveOccurred())
+
+	return values.Name(), func() { _ = os.Remove(values.Name()) }
+}
+
+var StrictValidationArgs = []string{
+	"--set", "gateway.validation.failurePolicy=Fail",
+	"--set", "gateway.validation.allowWarnings=false",
+	"--set", "gateway.validation.alwaysAcceptResources=false",
+}
+
+func RunAndCleanCommand(name string, arg ...string) []byte {
+	cmd := exec.Command(name, arg...)
+	b, err := cmd.Output()
+	// for debugging in Cloud Build
+	if err != nil {
+		if v, ok := err.(*exec.ExitError); ok {
+			fmt.Println("ExitError: ", string(v.Stderr))
+		}
+	}
+	Expect(err).To(BeNil())
+	cmd.Process.Kill()
+	cmd.Process.Release()
+	return b
+}
+
+func CheckGlooHealthyWithDeployments(testHelper *helper.SoloTestHelper, deploymentNames []string) {
+	for _, deploymentName := range deploymentNames {
+		RunAndCleanCommand("kubectl", "rollout", "status", "deployment", "-n", testHelper.InstallNamespace, deploymentName)
+	}
+	kube2e.GlooctlCheckEventuallyHealthy(2, testHelper, "90s")
+}
+
+func CheckGlooOssHealthy(testHelper *helper.SoloTestHelper) {
+	deploymentNames := []string{"gloo", "discovery", "gateway-proxy"}
+	CheckGlooHealthyWithDeployments(testHelper, deploymentNames)
+}
+
+func GetGlooServerVersion(ctx context.Context, namespace string) (v string) {
+	glooVersion, err := version.GetClientServerVersions(ctx, version.NewKube(namespace, ""))
+	Expect(err).To(BeNil())
+	Expect(len(glooVersion.GetServer())).To(Equal(1))
+	for _, container := range glooVersion.GetServer()[0].GetKubernetes().GetContainers() {
+		if v == "" {
+			v = container.OssTag
+		} else {
+			Expect(container.OssTag).To(Equal(v))
+		}
+	}
+	return v
 }
